Utworzenie tabeli ingredient, nutrient oraz tabeli nutrient-ingredient

+++ Project/diet/src/main/kotlin/pl/edu/pjwstk/s25236/diet_app/model/Nutrient.kt
@@ -1,3 +1,3 @@
 package pl.edu.pjwstk.s25236.diet_app.model
 
-class Nutrient(val id: Long = 0, val name: String) {}
+class Nutrient(val id: Short = 0, val name: String) {}
